+++ packages/component/src/ScrollToBottom/State2Context.js
@@ -3,14 +3,9 @@ import React from 'react';
 const context = React.createContext({
   animating: false,
   animatingToEnd: false,
-  atBottom: true,
-  atEnd: true,
-  atStart: false,
-  atTop: true,
-  mode: 'bottom',
   sticky: true
 });
 
-context.displayName = 'ScrollToBottomStateContext';
+context.displayName = 'ScrollToBottomState2Context';
 
 export default context;