+++ packages/component/src/hooks/internal/useStateContext.js
@@ -1,7 +1,8 @@
 import { useContext } from 'react';
 
-import StateContext from '../../ScrollToBottom/StateContext';
+import State1Context from '../../ScrollToBottom/State1Context';
+import State2Context from '../../ScrollToBottom/State2Context';
 
-export default function useStateContext() {
-  return useContext(StateContext);
+export default function useStateContext(tier) {
+  return useContext(tier === 2 ? State2Context : State1Context);
 }