+++ packages/component/src/hooks/useAnimating.js
@@ -1,7 +1,7 @@
 import useStateContext from './internal/useStateContext';
 
 export default function useAnimating() {
-  const { animating } = useStateContext();
+  const { animating } = useStateContext(2);
 
   return [animating];
 }