+++ packages/component/src/hooks/useAnimatingToEnd.js
@@ -1,7 +1,7 @@
 import useStateContext from './internal/useStateContext';
 
 export default function useAnimatingToEnd() {
-  const { animatingToEnd } = useStateContext();
+  const { animatingToEnd } = useStateContext(2);
 
   return [animatingToEnd];
 }