+++ packages/component/src/hooks/useAtEnd.js
@@ -1,7 +1,7 @@
 import useStateContext from './internal/useStateContext';
 
 export default function useAtEnd() {
-  const { atEnd } = useStateContext();
+  const { atEnd } = useStateContext(1);
 
   return [atEnd];
 }