+++ packages/component/src/hooks/useMode.js
@@ -1,7 +1,7 @@
 import useStateContext from './internal/useStateContext';
 
 export default function useMode() {
-  const { mode } = useStateContext();
+  const { mode } = useStateContext(1);
 
   return [mode];
 }