+++ packages/component/src/hooks/useSticky.js
@@ -1,7 +1,7 @@
 import useStateContext from './internal/useStateContext';
 
 export default function useSticky() {
-  const { sticky } = useStateContext();
+  const { sticky } = useStateContext(2);
 
   return [sticky];
 }