+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added `scroller` prop for limiting scroll distance when `mode` is set to `bottom`, in PR [#XXX](https://github.com/compulim/react-scroll-to-bottom/pull/XXX)
 - Added `initialScrollBehavior` prop for first scroll behavior. When set to `"auto"` (discrete scrolling), it will jump to end on initialization. in PR [#XXX](https://github.com/compulim/react-scroll-to-bottom/pull/XXX)
+- Added `debug` prop for dumping debug log to console, in PR [#XXX](https://github.com/compulim/react-scroll-to-bottom/pull/XXX)
+- Improved performance by separating `StateContext` into 2 tiers, in PR [#XXX](https://github.com/compulim/react-scroll-to-bottom/pull/XXX)
 
 ### Fixed